requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP Client
httpx==0.26.0

# Data Processing & Parsing
python-dateutil==2.8.2
unidecode==1.3.7

# Google Cloud
google-cloud-secret-manager==2.17.0
google-auth==2.26.2
gspread==6.0.0
google-auth-oauthlib==1.2.0

# Security
python-dotenv==1.0.0
cryptography==41.0.7

# Utilities
python-multipart==0.0.6

# Production Server
gunicorn==21.2.0
